Small changes to the example.
Those are the changes I had to make to the example to make it work on my own system.

`$timeout = 5;` could probably be moved to the server_check.php if you wish to keep the example smaller. And there is rarely ever a need to change this.
But I'd keep it in the example, as there is no way to change it before you include the check file. Unless you made the check into a function. But that would again complicate it more.